Merge Rails 4, Composer's and Zurb layouts
Composer has it's own application layout, which is useful. (It includes
navigation elements and flash notices already.) Zurb has it's own, which is
derived from the HTML 5 boilerplate, a bit. And Rails 4 has nothing particularly
special, except that it includes the data-turbolinks-track attribute on it's
stylesheet and javascript, which should be included as well.