Tom & I worked out a new database layout
The clock has a list of users and statues it cares about. The statuses are all
associated with the clock. Each status is associated with a user. Each status is
assigned an LCD index, which is exactly where on the clock face the status will
be written to. (Assuming the LCDs are labeled in Arduino 0-5.)